License Info and Functions Used in Manuscript
Changes to comments. Comments are included to reflect license information.
The versions of AP_calculations.m and vElectRx_FiberSpikeModel_GasserReconstruction.m
functions used to produce the figures in the manuscript are added.
The script compositeFigure.m is newly added.
Modified README.txt file.
Minor changes in runSimulationScript.m to reflect the figures in the manuscript.